+++ docs/changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Place any unreleased changes here, that are subject to release in coming versions :).
 
+* feat: Add profiles configuration option to Spring Boot.
 * fix: Fix bug in rabbitmq relation. When an application with units integrate with
        rabbitmq-server it can fail with RelationDataAccessError.